# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedGama Telehealth API",
...
        "contact": {
            "name": "MedGama Dev Team",
            "email": "dev@medgama.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Local Development"
        },
        {
            "url": "https://api.medgama.com/api",
            "description": "Production"
        }
    ],
    "paths": {
        "/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "List appointments (scoped by authenticated user role)",
                "operationId": "b8c7b364f8969997096bb5cc471278c2",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "confirmed",
                                "cancelled",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "patient_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated appointments (AppointmentResource)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create appointment (locks calendar slot atomically)",
                "operationId": "a115e378e7f96db9e8a6537cbdd95921",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "doctor_id",
                                    "appointment_type",
                                    "slot_id",
                                    "appointment_date",
                                    "appointment_time"
                                ],
                                "properties": {
                                    "doctor_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "patient_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "clinic_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "appointment_type": {
                                        "type": "string",
                                        "enum": [
                                            "inPerson",
                                            "online"
                                        ]
                                    },
                                    "slot_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "appointment_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "appointment_time": {
                                        "type": "string",
                                        "example": "14:30"
                                    },
                                    "confirmation_note": {
                                        "description": "Encrypted at rest",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment created (AppointmentResource)"
                    },
                    "422": {
                        "description": "Validation error / slot unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/appointments/{appointment}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Show appointment detail (HIPAA-audited)",
                "description": "Access is logged to health_data_audit_logs table. doctor_note and confirmation_note are encrypted at rest (AES-256-CBC).",
                "operationId": "a38fbb7f74ce31ce3a372d53c082515b",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment detail (AppointmentResource)"
                    },
                    "403": {
                        "description": "Not a participant",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update appointment (status, notes, video link)",
                "operationId": "2664db886fddb964fd77ccad385baff4",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "cancelled",
                                            "completed"
                                        ]
                                    },
                                    "doctor_note": {
                                        "description": "Encrypted at rest (AES-256-CBC)",
                                        "type": "string"
                                    },
                                    "confirmation_note": {
                                        "description": "Encrypted at rest",
                                        "type": "string"
                                    },
                                    "video_conference_link": {
                                        "type": "string",
                                        "format": "uri"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Soft-delete appointment (slot released, GDPR: pruned after 10 years)",
                "operationId": "340a1d1e4b71864ef4d72553cf13c68a",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment soft-deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "description": "╔══════════════════════════════════════════════════════════════════╗\n║  MedGama API — OpenAPI 3.0 Specification                       ║\n║  GDPR / KVKK / HIPAA Compliant Telehealth Platform             ║\n╚══════════════════════════════════════════════════════════════════╝",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "You are not authorized."
                    },
                    "code": {
                        "type": "string",
                        "example": "FORBIDDEN"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed."
                    },
                    "code": {
                        "type": "string",
                        "example": "VALIDATION_ERROR"
                    },
                    "errors": {
                        "type": "object",
                        "example": "{\"email\": [\"The email field is required.\"]}"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}